pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "hilti-vggt-runner"
version = "0.1.0"
description = "Runner utilities for Hilti MP4 and rosbag reconstruction with VGGT-SLAM."
requires-python = ">=3.11"
dependencies = [
  "matplotlib",
  "numpy",
  "open3d",
  "opencv-python",
  "PyYAML>=6.0",
  "scipy",
]

[project.optional-dependencies]
dev = ["pytest>=8.0"]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
